rename directory name
